requirements.txt
aiogram>=3.0.0
aiohttp>=3.8.5
python-dotenv>=1.0.0
matplotlib>=3.5.1
